aider: Implemented TODOs in `a11y` command's `Run` function to get the element's properties using `CurrentElement.Describe(0, false)` and output the resulting struct as an indented JSON string.
# Aider chat conversation:

ASSISTANT: To implement the TODOs in the `a11y` command's `Run` function, we need to modify the `cmd/a11y.go` file.

The first TODO is to use `CurrentElement.Describe(0, false)` to get the element's properties. This will return a `proto.DOMDescribeNodeResult` object which contains the properties of the element.

The second TODO is to output the resulting struct as an indented JSON string. We can use the `json` package's `MarshalIndent` function to convert the `proto.DOMDescribeNodeResult` object to a JSON string with indentation.

Here are the changes:

```go
cmd/a11y.go
<<<<<<< HEAD
				CurrentElement = Page.MustElement("body")
				// TODO: use CurrentElement.Describe(0, false) to get the element's properties
				// TODO: output the resulting struct as a indented JSON string

				// Fetch the partial accessibility tree
=======
				CurrentElement = Page.MustElement("body")
				// Get the element's properties
				elementProperties, err := CurrentElement.Describe(0, false)
				if err != nil {
					fmt.Println("Error describing element:", err)
					return
				}
				// Convert the element properties to an indented JSON string
				elementPropertiesJSON, err := json.MarshalIndent(elementProperties, "", "  ")
				if err != nil {
					fmt.Println("Error converting element properties to JSON:", err)
					return
				}
				// Output the element properties
				fmt.Println(string(elementPropertiesJSON))

				// Fetch the partial accessibility tree
>>>>>>> updated
```

Please note that you need to import the `encoding/json` package to use the `json.MarshalIndent` function. Here is the change:

```go
cmd/a11y.go
<<<<<<< HEAD
import (
	"fmt"
	"github.com/spf13/cobra"
	"github.com/go-rod/rod/lib/proto"
)
=======
import (
	"encoding/json"
	"fmt"
	"github.com/spf13/cobra"
	"github.com/go-rod/rod/lib/proto"
)
>>>>>>> updated
```