aider: Iterate over the Nodes of the partial tree and output relevant info, filtering out any nodes that have ignore set to true.
# Aider chat conversation:

ASSISTANT: To implement the TODOs and output relevant information about the tree, we need to modify the `cmd/a11y.go` file.

We will iterate over the Nodes of the partial tree and output relevant info, filtering out any nodes that have ignore set to true. The relevant info includes computed string as text, source, number of children and ids.

Here are the steps we will follow:

1. After fetching the partial accessibility tree, we will iterate over the nodes.
2. For each node, we will check if the `ignored` field is set to `true`. If it is, we will skip this node.
3. If the `ignored` field is not set to `true`, we will print the relevant information. This includes the computed string (if it exists), the source of the computed string, the number of children, and the ids of the children.

Here is the *edit block* for the changes:

```go
cmd/a11y.go
<<<<<<< HEAD
				//TODO: iterate over the Nodes of the partial tree and output relevant info
				// filter out any nodes that have ignore set to true
				// relevant info:
				// - computed string as text
				// - source
				// - numberof children and ids
				if !hasCurrentElement() {
					return
				}
				// ReportElement(CurrentElement)
=======
				// Iterate over the Nodes of the partial tree and output relevant info
				for _, node := range partialAXTree.Nodes {
					// Filter out any nodes that have ignore set to true
					if node.Ignored {
						continue
					}
					// Relevant info: computed string as text, source, number of children and ids
					fmt.Println("Node ID:", node.NodeID)
					if node.Name != nil {
						fmt.Println("Computed string:", node.Name.Value)
						for _, source := range node.Name.Sources {
							fmt.Println("Source:", source.Type)
						}
					}
					fmt.Println("Number of children:", len(node.ChildIds))
					fmt.Println("Child IDs:", node.ChildIds)
				}
				if !hasCurrentElement() {
					return
				}
				// ReportElement(CurrentElement)
>>>>>>> updated
```

Please note that this code assumes that the `Name` field and the `Sources` field within it are optional, as they are not present in all nodes in the example JSON. If they are guaranteed to be present in all nodes, the checks for `nil` can be removed.