aider: Added 'set' subcommand to set or toggle command-line flags.
# Aider chat conversation:

USER:
Please implement a new sub command 'set' that will set or toggle the various persistant commandline flags.
It will do so by setting the underlying global var to the value entered by the user.
In case of boolean vars, it will toggle the value from true to false or vice versa.
The command will always take at least one argument, the name of the commandline flag (e.g. it could be -v or --verbose)
The command will confirm by outputting the new value of the var

ASSISTANT: To implement the 'set' subcommand, we need to modify the `cmd/set.go` file. We will add a new `SetCmd` variable of type `*cobra.Command` that will handle the 'set' subcommand. This command will take at least one argument, the name of the command-line flag, and optionally a value to set the flag to. If no value is provided for a boolean flag, it will toggle the flag's current value. After setting the flag, the command will output the new value of the flag.

Here is the proposed change:

```go
cmd/set.go
<<<<<<< HEAD
=======
package cmd

import (
	"fmt"
	"github.com/spf13/cobra"
	"strconv"
)

var SetCmd = &cobra.Command{
	Use:   "set",
	Short: "Set or toggle the value of a command-line flag",
	Args:  cobra.MinimumNArgs(1),
	Run: func(cmd *cobra.Command, args []string) {
		flagName := args[0]
		flag := RootCmd.PersistentFlags().Lookup(flagName)
		if flag == nil {
			fmt.Println("Error: unknown flag:", flagName)
			return
		}
		var newValue string
		if len(args) > 1 {
			newValue = args[1]
		} else if flag.Value.Type() == "bool" {
			oldValue, _ := strconv.ParseBool(flag.Value.String())
			newValue = strconv.FormatBool(!oldValue)
		} else {
			fmt.Println("Error: no value provided for non-boolean flag:", flagName)
			return
		}
		if err := flag.Value.Set(newValue); err != nil {
			fmt.Println("Error setting flag:", err)
			return
		}
		fmt.Println("Set", flagName, "to", flag.Value)
	},
}

func init() {
	RootCmd.AddCommand(SetCmd)
}
>>>>>>> updated
```

This code creates a new 'set' command that takes at least one argument, the name of the flag to set. If a second argument is provided, it sets the flag to that value. If no second argument is provided and the flag is a boolean, it toggles the flag's value. If no second argument is provided and the flag is not a boolean, it outputs an error. After setting the flag, it outputs the new value of the flag.