aider: Added a new subcommand 'pick' to the RootCmd in the init function, which allows the user to pick a node by its id.
# Aider chat conversation:

ASSISTANT: To create a new subcommand 'pick', we need to add a new `cobra.Command` object to the `cmd` package. This command will take one parameter, a node-id, and will set the `CurrentElement` to the node that corresponds to this id. For now, we will stub out the actual setting of the `CurrentElement`.

Here is the change we need to make:

```go
cmd/a11y.go
<<<<<<< HEAD
func init() {
	RootCmd.AddCommand(quaxCmd)
}
=======
func init() {
	RootCmd.AddCommand(quaxCmd)
	RootCmd.AddCommand(pickCmd)
}

var pickCmd = &cobra.Command{
	Use:   "pick",
	Short: "Pick a node by its id",
	Run: func(cmd *cobra.Command, args []string) {
		if len(args) > 0 {
			nodeID := args[0]
			fmt.Println("Picking node with ID:", nodeID)
			// TODO: Set CurrentElement to the node that corresponds to this id
		} else {
			fmt.Println("Please provide a node ID")
		}
	},
}
>>>>>>> updated
```

This code adds a new command `pick` to the `RootCmd` in the `init` function. The `pick` command checks if a node ID is provided as an argument. If it is, it prints a message saying it's picking the node with that ID. If no node ID is provided, it prints a message asking for a node ID. The actual setting of the `CurrentElement` is stubbed out for now.