aider: Convert the `nodeID` from a string to an integer in the `pickCmd` function.
# Aider chat conversation:

USER: please see the todo marker about the nodeid needing to be an int

ASSISTANT: I see the TODO comment in the `pickCmd` function in the `cmd/a11y.go` file. It seems like the `nodeID` is currently being treated as a string, but it should be an integer.

To fix this, we need to convert the `nodeID` from a string to an integer using the `strconv.Atoi` function. This function will return an integer and an error. If the conversion is successful, the error will be `nil`, otherwise, it will contain information about what went wrong.

Here is the change that needs to be made:

```go
cmd/a11y.go
<<<<<<< HEAD
			nodeID := args[0]
			fmt.Println("Picking node with ID:", nodeID)
			// Set CurrentElement to the node that corresponds to this id
			Page.MustElementFromNode(&proto.DOMNode{NodeID: nodeID})
=======
			nodeID, err := strconv.Atoi(args[0])
			if err != nil {
				fmt.Println("Error converting node ID to integer:", err)
				return
			}
			fmt.Println("Picking node with ID:", nodeID)
			// Set CurrentElement to the node that corresponds to this id
			Page.MustElementFromNode(&proto.DOMNode{NodeID: proto.NodeID(nodeID)})
>>>>>>> updated
```

This change will ensure that the `nodeID` is treated as an integer, as required.