aider: Added import statements for `strconv` package in `cmd/a11y.go`.
# Aider chat conversation:

USER: I ran this command:

 go build .

And got this output:

# roderik/cmd
cmd/a11y.go:107:19: undefined: strconv
cmd/a11y.go:114:58: undefined: proto.NodeID

ASSISTANT: It seems like the `strconv` and `proto` packages are not imported in the `cmd/a11y.go` file.

To fix this, we need to add the import statements for these packages at the top of the file.

Here is the change that needs to be made:

```go
cmd/a11y.go
<<<<<<< HEAD
import (
	"encoding/json"
	"fmt"
	"github.com/go-rod/rod/lib/proto"
	"github.com/spf13/cobra"
)
=======
import (
	"encoding/json"
	"fmt"
	"strconv"
	"github.com/go-rod/rod/lib/proto"
	"github.com/spf13/cobra"
)
>>>>>>> updated
```

This change will ensure that the `strconv` and `proto` packages are imported and available for use in the `cmd/a11y.go` file.