aider: Change the type of `NodeID` to `proto.DOMNodeID` in `cmd/a11y.go` file.
# Aider chat conversation:

USER: I ran this command:

 go build .

And got this output:

# roderik/cmd
cmd/a11y.go:115:58: undefined: proto.NodeID

ASSISTANT: It seems like the `proto.NodeID` type is not defined in the `proto` package.

To fix this, we need to use the correct type for `NodeID`. According to the `github.com/go-rod/rod/lib/proto` package, the type for `NodeID` should be `proto.DOMNodeID`.

Here is the change that needs to be made:

```go
cmd/a11y.go
<<<<<<< HEAD
			// Set CurrentElement to the node that corresponds to this id
			Page.MustElementFromNode(&proto.DOMNode{NodeID: proto.NodeID(nodeID)})
=======
			// Set CurrentElement to the node that corresponds to this id
			Page.MustElementFromNode(&proto.DOMNode{NodeID: proto.DOMNodeID(nodeID)})
>>>>>>> updated
```

This change will ensure that the `NodeID` is of the correct type `proto.DOMNodeID`, as required.