aider: Output the Go type of the two ID variables `node.NodeID` and `node.BackendDOMNodeID` in the `quaxCmd` function.
# Aider chat conversation:

ASSISTANT: Sure, I see the TODO marker in the `quaxCmd` function. It's asking to output the Go type of the two ID variables `node.NodeID` and `node.BackendDOMNodeID`.

We can use the `reflect.TypeOf()` function from the `reflect` package to get the type of these variables.

Here is the change that needs to be made in the `cmd/a11y.go` file:

```go
cmd/a11y.go
<<<<<<< HEAD
				fmt.Println("Node ID:", node.NodeID)
				fmt.Println("Backend DOM Node ID:", node.BackendDOMNodeID)
=======
				fmt.Println("Node ID:", node.NodeID, "Type:", reflect.TypeOf(node.NodeID))
				fmt.Println("Backend DOM Node ID:", node.BackendDOMNodeID, "Type:", reflect.TypeOf(node.BackendDOMNodeID))
>>>>>>> updated
```

Please note that this change requires the `reflect` package. If it's not already imported, you need to add `import "reflect"` at the beginning of your file.