aider: Added new sub-command 'clear' with the alias 'cls' to clear the terminal screen using the appropriate command based on the operating system.
# Aider chat conversation:

USER: please implement a new sub cmmand called 'clear' with the alias of 'cls'
The command will use the unix command 'clear' to celar the screen.
If the app is running in a windows cmd shell, the command to execute is the cls equivalent.

ASSISTANT: To implement the new sub-command 'clear' with the alias 'cls', we need to modify the `cmd/root.go` file.

We will add a new command called `ClearCmd` and add it to the `RootCmd` in the `init` function. The `ClearCmd` will check the operating system and execute the appropriate command to clear the terminal.

Here are the changes:

```go
cmd/root.go
<<<<<<< HEAD
func init() {
	// Sub commands removed
	RootCmd.PersistentFlags().BoolVarP(&ShowNetActivity, "net-activity", "n", false, "Enable display of network events")
	RootCmd.PersistentFlags().BoolVarP(&Interactive, "interactive", "i", false, "Enable interactive mode")
	RootCmd.PersistentFlags().BoolVarP(&Verbose, "verbose", "v", false, "Enable verbose mode")
}
=======
func init() {
	// Sub commands removed
	RootCmd.PersistentFlags().BoolVarP(&ShowNetActivity, "net-activity", "n", false, "Enable display of network events")
	RootCmd.PersistentFlags().BoolVarP(&Interactive, "interactive", "i", false, "Enable interactive mode")
	RootCmd.PersistentFlags().BoolVarP(&Verbose, "verbose", "v", false, "Enable verbose mode")

	RootCmd.AddCommand(ClearCmd)
}
>>>>>>> updated
```

```go
cmd/root.go
<<<<<<< HEAD

=======
var ClearCmd = &cobra.Command{
	Use:     "clear",
	Aliases: []string{"cls"},
	Short:   "Clear the terminal screen",
	Long:    `This command will clear the terminal screen.`,
	Run: func(cmd *cobra.Command, args []string) {
		if runtime.GOOS == "windows" {
			cmd := exec.Command("cmd", "/c", "cls")
			cmd.Stdout = os.Stdout
			cmd.Run()
		} else {
			cmd := exec.Command("clear")
			cmd.Stdout = os.Stdout
			cmd.Run()
		}
	},
}
>>>>>>> updated
```

Please note that the `runtime` and `os/exec` packages are used in the new command. If these packages are not already imported at the top of the file, you will need to add them.