aider: Added new subcommand 'computedstyles' to cmd/css.go.
# Aider chat conversation:

USER: Please create a new sub command 'computedstyles' in the cmd/css.go file
Use the examples in inspect.go for a guideline of how to structure the sub comamnd

The computedstyles will output the computed styles of the curretly selected element in a format suitable for a .css file
Do not yet implement the actual comamnd logic, we will do that later.

ASSISTANT: To create a new subcommand 'computedstyles', we will need to modify the `cmd/css.go` file. We will create a new `cobra.Command` named `ComputedStyleCmd` and add it to the `init` function.

Here is the proposed change:

```go
cmd/css.go
<<<<<<< HEAD

=======
package cmd

import (
	"github.com/spf13/cobra"
)

var ComputedStyleCmd = &cobra.Command{
	Use:   "computedstyles",
	Short: "Output the computed styles of the currently selected element in a format suitable for a .css file",
	Run: func(cmd *cobra.Command, args []string) {
		// TODO: Implement command logic
	},
}

func init() {
	RootCmd.AddCommand(ComputedStyleCmd)
}
>>>>>>> updated
```

This code creates a new command `computedstyles` with a short description. The `Run` function is currently empty and will be filled with the actual command logic later. The `init` function adds the new command to the root command.