aider: Added classes `.three` and `.four` to specify the number of images in a row, with corresponding CSS rules for width.
# Aider chat conversation:

USER: I'd like to be able to specify the number of images I have in a row by addin adding an extra css class on the div. If this class is 'three', the current width setting of the img tags insode will apply Please also create a css rule that fits 4 images inside the container

ASSISTANT: To achieve this, we need to modify the CSS rules in the `assets/image-row.html` file. We will create two new classes, `.three` and `.four`, which will be used to specify the number of images in a row. The `.three` class will keep the current width setting, and the `.four` class will have a width of 25% to fit 4 images inside the container.

Here is the *edit block* for the changes:

```html
assets/image-row.html
<<<<<<< HEAD
   <style>
       .image-container {
           display: flex;
           justify-content: space-evenly;
           width: 100%;
       }
       .image-container img {
           width: 33.33%;
           height: auto;
       }
   </style>
=======
   <style>
       .image-container {
           display: flex;
           justify-content: space-evenly;
           width: 100%;
       }
       .image-container.three img {
           width: 33.33%;
           height: auto;
       }
       .image-container.four img {
           width: 25%;
           height: auto;
       }
   </style>
>>>>>>> updated
```

After this change, you can specify the number of images in a row by adding the `three` or `four` class to the `image-container` div. For example:

```html
<div class='image-container three'>
    <!-- three images here -->
</div>
<div class='image-container four'>
    <!-- four images here -->
</div>
```