aider: Added margin and adjusted width of images in `assets/image-row.html`.
# Aider chat conversation:

ASSISTANT: To achieve this, we need to modify the CSS in the `assets/image-row.html` file. We will add a margin to the images, and then use negative margins on the container to counteract the margins on the outer images.

Here is the change:

```html
assets/image-row.html
<<<<<<< HEAD
   <style>
       .image-container {
           display: flex;
           justify-content: space-evenly;
           width: 100%;
       }
       .image-container.three img {
           width: 33.33%;
           height: auto;
       }
       .image-container.four img {
           width: 25%;
           height: auto;
       }
   </style>
=======
   <style>
       .image-container {
           display: flex;
           justify-content: space-evenly;
           width: 100%;
           margin-left: -10px;
           margin-right: -10px;
       }
       .image-container img {
           margin-left: 10px;
           margin-right: 10px;
       }
       .image-container.three img {
           width: calc(33.33% - 20px);
           height: auto;
       }
       .image-container.four img {
           width: calc(25% - 20px);
           height: auto;
       }
   </style>
>>>>>>> updated
```

This change adds a 10px margin to the left and right of each image, and then subtracts that same amount from the container. The width of the images is also adjusted to account for the added margin. This will result in spacing between the images, but not on the outside of the first and last images in a row.